Refactor content relations table. #206
